install_shell_utilities: only run tasks if needed

+++ .dotfiles_meta/install_shell_utilities.fish
@@ -7,26 +7,40 @@ if not grep -q "$path_to_shell" /etc/shells
     chsh -s "$path_to_shell"
 end
 
-echo "Installing vim-plug for neovim"
-curl --create-dirs --location --output ~/.local/share/nvim/site/autoload/plug.vim https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+set -l vim_plug_path ~/.local/share/nvim/site/autoload/plug.vim
+if test ! -e "$vim_plug_path"
+    echo "Installing vim-plug for neovim"
+    curl --create-dirs --location --output "$vim_plug_path" https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+end
 
-echo "Installing the latest git-completion script for bash"
-curl --create-dirs --location --output ~/.config/bash/git-completion.bash https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+set -l bash_git_completion_path ~/.config/bash/git-completion.bash
+if test ! -e "$bash_git_completion_path"
+    echo "Installing the latest git-completion script for bash"
+    curl --create-dirs --location --output "$bash_git_completion_path" https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+end
 
-echo "Installing fisher and fisher plugins"
-curl -sL git.io/fisher | source && fisher install jorgebucaran/fisher
-fish -c "fisher update"
+if not functions --query fisher
+    echo "Installing fisher and fisher plugins"
+    curl -sL git.io/fisher | source && fisher install jorgebucaran/fisher
+    fish -c "fisher update"
+end
 
-echo "Configuring Tide prompt"
-tide configure --auto --style=Rainbow --prompt_colors='True color' --show_time=No --rainbow_prompt_separators=Angled --powerline_prompt_heads=Sharp --powerline_prompt_tails=Flat --powerline_prompt_style='One line' --prompt_spacing=Compact --icons='Few icons' --transient=Yes
-# only show a few items on right side of prompt to keep it short and unintrusive
-set tide_right_prompt_items status cmd_duration context
-tide reload # reload makes the changes take effect
+if test (count $tide_right_prompt_items) -gt 10
+    echo "Configuring Tide prompt"
+    tide configure --auto --style=Rainbow --prompt_colors='True color' --show_time=No --rainbow_prompt_separators=Angled --powerline_prompt_heads=Sharp --powerline_prompt_tails=Flat --powerline_prompt_style='One line' --prompt_spacing=Compact --icons='Few icons' --transient=Yes
+    # only show a few items on right side of prompt to keep it short and unintrusive
+    set tide_right_prompt_items status cmd_duration context
+    tide reload # reload makes the changes take effect
+end
 
-echo "Installing plugins for nnn"
-curl -Ls https://raw.githubusercontent.com/jarun/nnn/master/plugins/getplugs | sh
+if ! test -e ~/.config/nnn/plugins
+    echo "Installing plugins for nnn"
+    curl -Ls https://raw.githubusercontent.com/jarun/nnn/master/plugins/getplugs | sh
+end
 
-# Remove semibold variant so no editor will automatically pick it for bold
-# because it is almost indistinguishable from the regular variant
-echo "Removing JetBrains Mono SemiBold fonts"
-rm -f ~/Library/Fonts/JetBrainsMono{,NL}-{SemiBold, SemiBoldItalic}.ttf
+if test -e ~/Library/Fonts/JetBrainsMono-SemiBold.ttf
+    # Remove semibold variant so no editor will automatically pick it for bold
+    # because it is almost indistinguishable from the regular variant
+    echo "Removing JetBrains Mono SemiBold fonts"
+    rm -f ~/Library/Fonts/JetBrainsMono{,NL}-{SemiBold, SemiBoldItalic}.ttf
+end
